Add an event for uploaded files and a command to replace the source (#6366)
* Add an event for the creation of a file

This cannot be placed in `FileAction::create()` because the physical file isn’t present at that time.

* Add a command to replace the source of a file

* Add an option to override the filename

This is useful when replacing the file with a temporarily generated name.

* Rename the event and only fire it for uploaded files

The method `createFromExistingFile()` is a programmatic call that usually occurs during a data import. Any additional actions, such as the modification of images, should only be carried out during an upload or when rebuilding data.

* Clean up the old file when replacing an uploaded file